requirements.txt
transformers
torch
torchvision
gradio
Pillow
qwen-vl-utils
accelerate
byaldi